refactor(android): extract AndroidConsentController and AndroidAdDiagnostics to their own files
iOS gets IosConsentController.kt, IosAdDiagnostics.kt, IosTrackingAuthorization.kt
as top-level files. Android buried the equivalent ~215-line consent controller and
~55-line diagnostics class as private classes at the bottom of the 904-line
AndroidGoogleAdManager.kt. The code comments explicitly say the two consent
controllers are "kept identical on purpose" — yet they lived in structurally
different places, making the natural review move (diffing them against each
other) harder than it should be.

Pure move: both classes go from `private` to `internal` (file-private no longer
applies) with no other behavior change. AndroidGoogleAdManager.kt drops from 904
to 621 lines and is now line-for-line diffable against IosGoogleAdManager.kt.
Both extracted classes are internal and absent from admob-cmp-core.klib.api, so
this is ABI-neutral (verified: checkKotlinAbi produces no dump diff).

Re-keys the three detekt-baseline.xml entries whose ID embeds the old file name
(AndroidGoogleAdManager.kt$AndroidConsentController$...) to the new file
(AndroidConsentController.kt$...) — the baseline key is the literal file path,
so the move would otherwise silently stop matching and reintroduce those
findings as new failures.

Also drops the dangling "Task 1 of this plan" reference from
AndroidAdDiagnostics.captureSnapshotOnMain's KDoc while moving it — the plan
document it pointed at was never committed to this repo.
