+++ admob-cmp-core/detekt-baseline.xml
@@ -2,7 +2,7 @@
 <SmellBaseline>
   <ManuallySuppressedIssues/>
   <CurrentIssues>
-    <ID>ComplexCondition:AndroidGoogleAdManager.kt$AndroidConsentController$config.testMode &amp;&amp; (config.testDeviceIds.isNotEmpty() || config.debugOptions.consentTestDeviceIds.isNotEmpty() || config.debugGeography != ConsentDebugGeography.Disabled)</ID>
+    <ID>ComplexCondition:AndroidConsentController.kt$AndroidConsentController$config.testMode &amp;&amp; (config.testDeviceIds.isNotEmpty() || config.debugOptions.consentTestDeviceIds.isNotEmpty() || config.debugGeography != ConsentDebugGeography.Disabled)</ID>
     <ID>ComplexCondition:FullScreenSlotCore.kt$FullScreenSlotCore$blockedError == null &amp;&amp; activePresentation.load() == null &amp;&amp; fresh.isNotEmpty() &amp;&amp; presentError == null</ID>
     <ID>ComplexCondition:NativeAdCoordinatorCore.kt$NativeAdCoordinatorCore$entry.sessionKey != sessionKey || entry.slotKey != slotKey || entry.generation != holder.core.slotGenerationFor(slotKey) || entry.placement != placement</ID>
     <ID>ComplexCondition:NativeAdCoordinatorCore.kt$NativeAdCoordinatorCore$entry.sessionKey != sessionKey || entry.slotKey != slotKey || entry.placement != placement || entry.rendererId != rendererId</ID>
@@ -20,8 +20,8 @@
     <ID>MaxLineLength:AndroidFullScreenSlots.kt$&lt;no name provided>$override fun onAdFailedToShowFullScreenContent(error: FullScreenContentError)</ID>
     <ID>MaxLineLength:AndroidFullScreenSlots.kt$AndroidRewardedInterstitialSlot$RewardedInterstitialAd</ID>
     <ID>MaxLineLength:AndroidFullScreenSlots.kt$com.google.android.libraries.ads.mobile.sdk.rewarded.ServerSideVerificationOptions(it.userId.orEmpty(), it.customData.orEmpty())</ID>
-    <ID>MaxLineLength:AndroidGoogleAdManager.kt$AndroidConsentController$if</ID>
-    <ID>MaxLineLength:AndroidGoogleAdManager.kt$AndroidConsentController${ if (continuation.isActive) continuation.resume(AdError(code = it.errorCode.toString(), message = it.message)) }</ID>
+    <ID>MaxLineLength:AndroidConsentController.kt$AndroidConsentController$if</ID>
+    <ID>MaxLineLength:AndroidConsentController.kt$AndroidConsentController${ if (continuation.isActive) continuation.resume(AdError(code = it.errorCode.toString(), message = it.message)) }</ID>
     <ID>MaxLineLength:AndroidGoogleAdManager.kt$AndroidGoogleAdManager$slots.getOrPut(AdSlotKey(ownedPlacement.id, ownedPlacement.format)) { AndroidAppOpenSlot(ownedPlacement, activityProvider, _events, ::adRequestBlockedError, ::onPresentationChanged, fullScreenArbiter, fullScreenAudioController) } as AppOpenAdController</ID>
     <ID>MaxLineLength:AndroidGoogleAdManager.kt$AndroidGoogleAdManager$slots.getOrPut(AdSlotKey(ownedPlacement.id, ownedPlacement.format)) { AndroidInterstitialSlot(ownedPlacement, activityProvider, _events, ::adRequestBlockedError, ::onPresentationChanged, fullScreenArbiter, fullScreenAudioController) } as InterstitialAdController</ID>
     <ID>MaxLineLength:AndroidGoogleAdManager.kt$AndroidGoogleAdManager$slots.getOrPut(AdSlotKey(ownedPlacement.id, ownedPlacement.format)) { AndroidRewardedInterstitialSlot(ownedPlacement, activityProvider, _events, ::adRequestBlockedError, ::onPresentationChanged, fullScreenArbiter, fullScreenAudioController) } as RewardedInterstitialAdController</ID>
